refactor RadialProgressThree to ProgressRadial
